requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
paho-mqtt==1.5.0
sensor==6
smbus==1.1.post2
Werkzeug==1.0.0
